Modify existing client classes to be abstract base classes
