+++ java-client/src/main/java/org/opensearch/client/opensearch/OpenSearchAsyncClientBase.java
@@ -33,13 +33,14 @@
 package org.opensearch.client.opensearch;
 
 import java.io.IOException;
+import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 import javax.annotation.Nullable;
 import org.opensearch.client.ApiClient;
 import org.opensearch.client.opensearch._types.ErrorResponse;
 import org.opensearch.client.opensearch._types.OpenSearchException;
-import org.opensearch.client.opensearch.cat.OpenSearchCatClient;
-import org.opensearch.client.opensearch.cluster.OpenSearchClusterClient;
+import org.opensearch.client.opensearch.cat.OpenSearchCatAsyncClient;
+import org.opensearch.client.opensearch.cluster.OpenSearchClusterAsyncClient;
 import org.opensearch.client.opensearch.core.BulkRequest;
 import org.opensearch.client.opensearch.core.BulkResponse;
 import org.opensearch.client.opensearch.core.ClearScrollRequest;
@@ -121,15 +122,14 @@ import org.opensearch.client.opensearch.core.pit.DeletePitRequest;
 import org.opensearch.client.opensearch.core.pit.DeletePitResponse;
 import org.opensearch.client.opensearch.core.pit.ListAllPitRequest;
 import org.opensearch.client.opensearch.core.pit.ListAllPitResponse;
-import org.opensearch.client.opensearch.dangling_indices.OpenSearchDanglingIndicesClient;
-import org.opensearch.client.opensearch.features.OpenSearchFeaturesClient;
-import org.opensearch.client.opensearch.generic.OpenSearchGenericClient;
-import org.opensearch.client.opensearch.indices.OpenSearchIndicesClient;
-import org.opensearch.client.opensearch.ingest.OpenSearchIngestClient;
-import org.opensearch.client.opensearch.nodes.OpenSearchNodesClient;
-import org.opensearch.client.opensearch.shutdown.OpenSearchShutdownClient;
-import org.opensearch.client.opensearch.snapshot.OpenSearchSnapshotClient;
-import org.opensearch.client.opensearch.tasks.OpenSearchTasksClient;
+import org.opensearch.client.opensearch.dangling_indices.OpenSearchDanglingIndicesAsyncClient;
+import org.opensearch.client.opensearch.features.OpenSearchFeaturesAsyncClient;
+import org.opensearch.client.opensearch.indices.OpenSearchIndicesAsyncClient;
+import org.opensearch.client.opensearch.ingest.OpenSearchIngestAsyncClient;
+import org.opensearch.client.opensearch.nodes.OpenSearchNodesAsyncClient;
+import org.opensearch.client.opensearch.shutdown.OpenSearchShutdownAsyncClient;
+import org.opensearch.client.opensearch.snapshot.OpenSearchSnapshotAsyncClient;
+import org.opensearch.client.opensearch.tasks.OpenSearchTasksAsyncClient;
 import org.opensearch.client.transport.JsonEndpoint;
 import org.opensearch.client.transport.OpenSearchTransport;
 import org.opensearch.client.transport.TransportOptions;
@@ -140,64 +140,53 @@ import org.opensearch.client.util.ObjectBuilder;
 /**
  * Client for the namespace.
  */
-public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchClient> {
+public abstract class OpenSearchAsyncClientBase<Self extends OpenSearchAsyncClientBase<Self>>
+        extends ApiClient<OpenSearchTransport, Self> {
 
-    public OpenSearchClient(OpenSearchTransport transport) {
-        super(transport, null);
-    }
-
-    public OpenSearchClient(OpenSearchTransport transport, @Nullable TransportOptions transportOptions) {
+    public OpenSearchAsyncClientBase(OpenSearchTransport transport, @Nullable TransportOptions transportOptions) {
         super(transport, transportOptions);
     }
 
-    @Override
-    public OpenSearchClient withTransportOptions(@Nullable TransportOptions transportOptions) {
-        return new OpenSearchClient(this.transport, transportOptions);
-    }
-
     // ----- Child clients
-    public OpenSearchGenericClient generic() {
-        return new OpenSearchGenericClient(this.transport, this.transportOptions);
+
+    public OpenSearchCatAsyncClient cat() {
+        return new OpenSearchCatAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchCatClient cat() {
-        return new OpenSearchCatClient(this.transport, this.transportOptions);
+    public OpenSearchClusterAsyncClient cluster() {
+        return new OpenSearchClusterAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchClusterClient cluster() {
-        return new OpenSearchClusterClient(this.transport, this.transportOptions);
+    public OpenSearchDanglingIndicesAsyncClient danglingIndices() {
+        return new OpenSearchDanglingIndicesAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchDanglingIndicesClient danglingIndices() {
-        return new OpenSearchDanglingIndicesClient(this.transport, this.transportOptions);
+    public OpenSearchFeaturesAsyncClient features() {
+        return new OpenSearchFeaturesAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchFeaturesClient features() {
-        return new OpenSearchFeaturesClient(this.transport, this.transportOptions);
+    public OpenSearchIndicesAsyncClient indices() {
+        return new OpenSearchIndicesAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchIndicesClient indices() {
-        return new OpenSearchIndicesClient(this.transport, this.transportOptions);
+    public OpenSearchIngestAsyncClient ingest() {
+        return new OpenSearchIngestAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchIngestClient ingest() {
-        return new OpenSearchIngestClient(this.transport, this.transportOptions);
+    public OpenSearchNodesAsyncClient nodes() {
+        return new OpenSearchNodesAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchNodesClient nodes() {
-        return new OpenSearchNodesClient(this.transport, this.transportOptions);
+    public OpenSearchShutdownAsyncClient shutdown() {
+        return new OpenSearchShutdownAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchShutdownClient shutdown() {
-        return new OpenSearchShutdownClient(this.transport, this.transportOptions);
+    public OpenSearchSnapshotAsyncClient snapshot() {
+        return new OpenSearchSnapshotAsyncClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchSnapshotClient snapshot() {
-        return new OpenSearchSnapshotClient(this.transport, this.transportOptions);
-    }
-
-    public OpenSearchTasksClient tasks() {
-        return new OpenSearchTasksClient(this.transport, this.transportOptions);
+    public OpenSearchTasksAsyncClient tasks() {
+        return new OpenSearchTasksAsyncClient(this.transport, this.transportOptions);
     }
 
     // ----- Endpoint: bulk
@@ -209,14 +198,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public BulkResponse bulk(BulkRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<BulkResponse> bulk(BulkRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<BulkRequest, BulkResponse, ErrorResponse> endpoint = (JsonEndpoint<
             BulkRequest,
             BulkResponse,
             ErrorResponse>) BulkRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -229,7 +218,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final BulkResponse bulk(Function<BulkRequest.Builder, ObjectBuilder<BulkRequest>> fn) throws IOException, OpenSearchException {
+    public final CompletableFuture<BulkResponse> bulk(Function<BulkRequest.Builder, ObjectBuilder<BulkRequest>> fn) throws IOException,
+        OpenSearchException {
         return bulk(fn.apply(new BulkRequest.Builder()).build());
     }
 
@@ -240,8 +230,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public BulkResponse bulk() throws IOException, OpenSearchException {
-        return this.transport.performRequest(new BulkRequest.Builder().build(), BulkRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<BulkResponse> bulk() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(new BulkRequest.Builder().build(), BulkRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: clear_scroll
@@ -252,14 +242,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public ClearScrollResponse clearScroll(ClearScrollRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<ClearScrollResponse> clearScroll(ClearScrollRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ClearScrollRequest, ClearScrollResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ClearScrollRequest,
             ClearScrollResponse,
             ErrorResponse>) ClearScrollRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -271,8 +261,9 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final ClearScrollResponse clearScroll(Function<ClearScrollRequest.Builder, ObjectBuilder<ClearScrollRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<ClearScrollResponse> clearScroll(
+        Function<ClearScrollRequest.Builder, ObjectBuilder<ClearScrollRequest>> fn
+    ) throws IOException, OpenSearchException {
         return clearScroll(fn.apply(new ClearScrollRequest.Builder()).build());
     }
 
@@ -282,8 +273,12 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public ClearScrollResponse clearScroll() throws IOException, OpenSearchException {
-        return this.transport.performRequest(new ClearScrollRequest.Builder().build(), ClearScrollRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<ClearScrollResponse> clearScroll() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
+            new ClearScrollRequest.Builder().build(),
+            ClearScrollRequest._ENDPOINT,
+            this.transportOptions
+        );
     }
 
     // ----- Endpoint: count
@@ -294,14 +289,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public CountResponse count(CountRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<CountResponse> count(CountRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<CountRequest, CountResponse, ErrorResponse> endpoint = (JsonEndpoint<
             CountRequest,
             CountResponse,
             ErrorResponse>) CountRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -313,7 +308,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final CountResponse count(Function<CountRequest.Builder, ObjectBuilder<CountRequest>> fn) throws IOException,
+    public final CompletableFuture<CountResponse> count(Function<CountRequest.Builder, ObjectBuilder<CountRequest>> fn) throws IOException,
         OpenSearchException {
         return count(fn.apply(new CountRequest.Builder()).build());
     }
@@ -324,8 +319,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public CountResponse count() throws IOException, OpenSearchException {
-        return this.transport.performRequest(new CountRequest.Builder().build(), CountRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<CountResponse> count() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(new CountRequest.Builder().build(), CountRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: create
@@ -339,14 +334,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument> CreateResponse create(CreateRequest<TDocument> request) throws IOException, OpenSearchException {
+    public <TDocument> CompletableFuture<CreateResponse> create(CreateRequest<TDocument> request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<CreateRequest<?>, CreateResponse, ErrorResponse> endpoint = (JsonEndpoint<
             CreateRequest<?>,
             CreateResponse,
             ErrorResponse>) CreateRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -361,8 +356,9 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TDocument> CreateResponse create(Function<CreateRequest.Builder<TDocument>, ObjectBuilder<CreateRequest<TDocument>>> fn)
-        throws IOException, OpenSearchException {
+    public final <TDocument> CompletableFuture<CreateResponse> create(
+        Function<CreateRequest.Builder<TDocument>, ObjectBuilder<CreateRequest<TDocument>>> fn
+    ) throws IOException, OpenSearchException {
         return create(fn.apply(new CreateRequest.Builder<TDocument>()).build());
     }
 
@@ -375,14 +371,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public CreatePitResponse createPit(CreatePitRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<CreatePitResponse> createPit(CreatePitRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<CreatePitRequest, CreatePitResponse, ErrorResponse> endpoint = (JsonEndpoint<
             CreatePitRequest,
             CreatePitResponse,
             ErrorResponse>) CreatePitRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -395,8 +391,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final CreatePitResponse createPit(Function<CreatePitRequest.Builder, ObjectBuilder<CreatePitRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<CreatePitResponse> createPit(Function<CreatePitRequest.Builder, ObjectBuilder<CreatePitRequest>> fn)
+        throws IOException, OpenSearchException {
         return createPit(fn.apply(new CreatePitRequest.Builder()).build());
     }
 
@@ -408,14 +404,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public DeleteResponse delete(DeleteRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<DeleteResponse> delete(DeleteRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeleteRequest, DeleteResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeleteRequest,
             DeleteResponse,
             ErrorResponse>) DeleteRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -427,8 +423,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final DeleteResponse delete(Function<DeleteRequest.Builder, ObjectBuilder<DeleteRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<DeleteResponse> delete(Function<DeleteRequest.Builder, ObjectBuilder<DeleteRequest>> fn)
+        throws IOException, OpenSearchException {
         return delete(fn.apply(new DeleteRequest.Builder()).build());
     }
 
@@ -440,14 +436,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public DeletePitResponse deletePit(DeletePitRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<DeletePitResponse> deletePit(DeletePitRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeletePitRequest, DeletePitResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeletePitRequest,
             DeletePitResponse,
             ErrorResponse>) DeletePitRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -459,8 +455,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final DeletePitResponse deletePit(Function<DeletePitRequest.Builder, ObjectBuilder<DeletePitRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<DeletePitResponse> deletePit(Function<DeletePitRequest.Builder, ObjectBuilder<DeletePitRequest>> fn)
+        throws IOException, OpenSearchException {
         return deletePit(fn.apply(new DeletePitRequest.Builder()).build());
     }
 
@@ -472,14 +468,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public DeleteByQueryResponse deleteByQuery(DeleteByQueryRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<DeleteByQueryResponse> deleteByQuery(DeleteByQueryRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeleteByQueryRequest, DeleteByQueryResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeleteByQueryRequest,
             DeleteByQueryResponse,
             ErrorResponse>) DeleteByQueryRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -491,8 +487,9 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final DeleteByQueryResponse deleteByQuery(Function<DeleteByQueryRequest.Builder, ObjectBuilder<DeleteByQueryRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<DeleteByQueryResponse> deleteByQuery(
+        Function<DeleteByQueryRequest.Builder, ObjectBuilder<DeleteByQueryRequest>> fn
+    ) throws IOException, OpenSearchException {
         return deleteByQuery(fn.apply(new DeleteByQueryRequest.Builder()).build());
     }
 
@@ -505,15 +502,15 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public DeleteByQueryRethrottleResponse deleteByQueryRethrottle(DeleteByQueryRethrottleRequest request) throws IOException,
-        OpenSearchException {
+    public CompletableFuture<DeleteByQueryRethrottleResponse> deleteByQueryRethrottle(DeleteByQueryRethrottleRequest request)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeleteByQueryRethrottleRequest, DeleteByQueryRethrottleResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeleteByQueryRethrottleRequest,
             DeleteByQueryRethrottleResponse,
             ErrorResponse>) DeleteByQueryRethrottleRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -526,7 +523,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final DeleteByQueryRethrottleResponse deleteByQueryRethrottle(
+    public final CompletableFuture<DeleteByQueryRethrottleResponse> deleteByQueryRethrottle(
         Function<DeleteByQueryRethrottleRequest.Builder, ObjectBuilder<DeleteByQueryRethrottleRequest>> fn
     ) throws IOException, OpenSearchException {
         return deleteByQueryRethrottle(fn.apply(new DeleteByQueryRethrottleRequest.Builder()).build());
@@ -540,14 +537,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public DeleteScriptResponse deleteScript(DeleteScriptRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<DeleteScriptResponse> deleteScript(DeleteScriptRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeleteScriptRequest, DeleteScriptResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeleteScriptRequest,
             DeleteScriptResponse,
             ErrorResponse>) DeleteScriptRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -559,8 +556,9 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final DeleteScriptResponse deleteScript(Function<DeleteScriptRequest.Builder, ObjectBuilder<DeleteScriptRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<DeleteScriptResponse> deleteScript(
+        Function<DeleteScriptRequest.Builder, ObjectBuilder<DeleteScriptRequest>> fn
+    ) throws IOException, OpenSearchException {
         return deleteScript(fn.apply(new DeleteScriptRequest.Builder()).build());
     }
 
@@ -572,14 +570,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public BooleanResponse exists(ExistsRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<BooleanResponse> exists(ExistsRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ExistsRequest, BooleanResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ExistsRequest,
             BooleanResponse,
             ErrorResponse>) ExistsRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -591,8 +589,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final BooleanResponse exists(Function<ExistsRequest.Builder, ObjectBuilder<ExistsRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<BooleanResponse> exists(Function<ExistsRequest.Builder, ObjectBuilder<ExistsRequest>> fn)
+        throws IOException, OpenSearchException {
         return exists(fn.apply(new ExistsRequest.Builder()).build());
     }
 
@@ -604,14 +602,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public BooleanResponse existsSource(ExistsSourceRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<BooleanResponse> existsSource(ExistsSourceRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ExistsSourceRequest, BooleanResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ExistsSourceRequest,
             BooleanResponse,
             ErrorResponse>) ExistsSourceRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -623,8 +621,9 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final BooleanResponse existsSource(Function<ExistsSourceRequest.Builder, ObjectBuilder<ExistsSourceRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<BooleanResponse> existsSource(
+        Function<ExistsSourceRequest.Builder, ObjectBuilder<ExistsSourceRequest>> fn
+    ) throws IOException, OpenSearchException {
         return existsSource(fn.apply(new ExistsSourceRequest.Builder()).build());
     }
 
@@ -636,8 +635,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument> ExplainResponse<TDocument> explain(ExplainRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<ExplainResponse<TDocument>> explain(ExplainRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ExplainRequest, ExplainResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             ExplainRequest,
@@ -649,7 +648,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -661,7 +660,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TDocument> ExplainResponse<TDocument> explain(
+    public final <TDocument> CompletableFuture<ExplainResponse<TDocument>> explain(
         Function<ExplainRequest.Builder, ObjectBuilder<ExplainRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -673,18 +672,16 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
     /**
      * Returns the information about the capabilities of fields among multiple
      * indices.
-     *
-     *
      */
 
-    public FieldCapsResponse fieldCaps(FieldCapsRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<FieldCapsResponse> fieldCaps(FieldCapsRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<FieldCapsRequest, FieldCapsResponse, ErrorResponse> endpoint = (JsonEndpoint<
             FieldCapsRequest,
             FieldCapsResponse,
             ErrorResponse>) FieldCapsRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -697,8 +694,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final FieldCapsResponse fieldCaps(Function<FieldCapsRequest.Builder, ObjectBuilder<FieldCapsRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<FieldCapsResponse> fieldCaps(Function<FieldCapsRequest.Builder, ObjectBuilder<FieldCapsRequest>> fn)
+        throws IOException, OpenSearchException {
         return fieldCaps(fn.apply(new FieldCapsRequest.Builder()).build());
     }
 
@@ -709,8 +706,12 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public FieldCapsResponse fieldCaps() throws IOException, OpenSearchException {
-        return this.transport.performRequest(new FieldCapsRequest.Builder().build(), FieldCapsRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<FieldCapsResponse> fieldCaps() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
+            new FieldCapsRequest.Builder().build(),
+            FieldCapsRequest._ENDPOINT,
+            this.transportOptions
+        );
     }
 
     // ----- Endpoint: get
@@ -721,8 +722,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument> GetResponse<TDocument> get(GetRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<GetResponse<TDocument>> get(GetRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<GetRequest, GetResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             GetRequest,
@@ -734,7 +735,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -746,7 +747,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TDocument> GetResponse<TDocument> get(
+    public final <TDocument> CompletableFuture<GetResponse<TDocument>> get(
         Function<GetRequest.Builder, ObjectBuilder<GetRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -761,14 +762,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public GetScriptResponse getScript(GetScriptRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<GetScriptResponse> getScript(GetScriptRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<GetScriptRequest, GetScriptResponse, ErrorResponse> endpoint = (JsonEndpoint<
             GetScriptRequest,
             GetScriptResponse,
             ErrorResponse>) GetScriptRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -780,8 +781,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final GetScriptResponse getScript(Function<GetScriptRequest.Builder, ObjectBuilder<GetScriptRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<GetScriptResponse> getScript(Function<GetScriptRequest.Builder, ObjectBuilder<GetScriptRequest>> fn)
+        throws IOException, OpenSearchException {
         return getScript(fn.apply(new GetScriptRequest.Builder()).build());
     }
 
@@ -792,8 +793,12 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      *
      */
-    public GetScriptContextResponse getScriptContext() throws IOException, OpenSearchException {
-        return this.transport.performRequest(GetScriptContextRequest._INSTANCE, GetScriptContextRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<GetScriptContextResponse> getScriptContext() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
+            GetScriptContextRequest._INSTANCE,
+            GetScriptContextRequest._ENDPOINT,
+            this.transportOptions
+        );
     }
 
     // ----- Endpoint: get_script_languages
@@ -803,8 +808,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      *
      */
-    public GetScriptLanguagesResponse getScriptLanguages() throws IOException, OpenSearchException {
-        return this.transport.performRequest(
+    public CompletableFuture<GetScriptLanguagesResponse> getScriptLanguages() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
             GetScriptLanguagesRequest._INSTANCE,
             GetScriptLanguagesRequest._ENDPOINT,
             this.transportOptions
@@ -819,8 +824,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument> GetSourceResponse<TDocument> getSource(GetSourceRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<GetSourceResponse<TDocument>> getSource(GetSourceRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<GetSourceRequest, GetSourceResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             GetSourceRequest,
@@ -832,7 +837,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -844,7 +849,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TDocument> GetSourceResponse<TDocument> getSource(
+    public final <TDocument> CompletableFuture<GetSourceResponse<TDocument>> getSource(
         Function<GetSourceRequest.Builder, ObjectBuilder<GetSourceRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -859,14 +864,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument> IndexResponse index(IndexRequest<TDocument> request) throws IOException, OpenSearchException {
+    public <TDocument> CompletableFuture<IndexResponse> index(IndexRequest<TDocument> request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<IndexRequest<?>, IndexResponse, ErrorResponse> endpoint = (JsonEndpoint<
             IndexRequest<?>,
             IndexResponse,
             ErrorResponse>) IndexRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -878,20 +883,10 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TDocument> IndexResponse index(Function<IndexRequest.Builder<TDocument>, ObjectBuilder<IndexRequest<TDocument>>> fn)
-        throws IOException, OpenSearchException {
+    public final <TDocument> CompletableFuture<IndexResponse> index(
+        Function<IndexRequest.Builder<TDocument>, ObjectBuilder<IndexRequest<TDocument>>> fn
+    ) throws IOException, OpenSearchException {
         return index(fn.apply(new IndexRequest.Builder<TDocument>()).build());
-    }
-
-    // ----- Endpoint: info
-
-    /**
-     * Returns basic information about the cluster.
-     *
-     *
-     */
-    public InfoResponse info() throws IOException, OpenSearchException {
-        return this.transport.performRequest(InfoRequest._INSTANCE, InfoRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: list_point_in_time
@@ -902,8 +897,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public ListAllPitResponse listAllPit() throws IOException, OpenSearchException {
-        return this.transport.performRequest(ListAllPitRequest._INSTANCE, ListAllPitRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<ListAllPitResponse> listAllPit() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(ListAllPitRequest._INSTANCE, ListAllPitRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: mget
@@ -914,8 +909,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument> MgetResponse<TDocument> mget(MgetRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<MgetResponse<TDocument>> mget(MgetRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<MgetRequest, MgetResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             MgetRequest,
@@ -927,7 +922,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -939,7 +934,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TDocument> MgetResponse<TDocument> mget(
+    public final <TDocument> CompletableFuture<MgetResponse<TDocument>> mget(
         Function<MgetRequest.Builder, ObjectBuilder<MgetRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -954,8 +949,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument> MsearchResponse<TDocument> msearch(MsearchRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<MsearchResponse<TDocument>> msearch(MsearchRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<MsearchRequest, MsearchResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             MsearchRequest,
@@ -967,7 +962,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -979,7 +974,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TDocument> MsearchResponse<TDocument> msearch(
+    public final <TDocument> CompletableFuture<MsearchResponse<TDocument>> msearch(
         Function<MsearchRequest.Builder, ObjectBuilder<MsearchRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -994,8 +989,10 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument> MsearchTemplateResponse<TDocument> msearchTemplate(MsearchTemplateRequest request, Class<TDocument> tDocumentClass)
-        throws IOException, OpenSearchException {
+    public <TDocument> CompletableFuture<MsearchTemplateResponse<TDocument>> msearchTemplate(
+        MsearchTemplateRequest request,
+        Class<TDocument> tDocumentClass
+    ) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<MsearchTemplateRequest, MsearchTemplateResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             MsearchTemplateRequest,
@@ -1007,7 +1004,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1019,7 +1016,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TDocument> MsearchTemplateResponse<TDocument> msearchTemplate(
+    public final <TDocument> CompletableFuture<MsearchTemplateResponse<TDocument>> msearchTemplate(
         Function<MsearchTemplateRequest.Builder, ObjectBuilder<MsearchTemplateRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1034,14 +1031,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public MtermvectorsResponse mtermvectors(MtermvectorsRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<MtermvectorsResponse> mtermvectors(MtermvectorsRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<MtermvectorsRequest, MtermvectorsResponse, ErrorResponse> endpoint = (JsonEndpoint<
             MtermvectorsRequest,
             MtermvectorsResponse,
             ErrorResponse>) MtermvectorsRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1053,8 +1050,9 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final MtermvectorsResponse mtermvectors(Function<MtermvectorsRequest.Builder, ObjectBuilder<MtermvectorsRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<MtermvectorsResponse> mtermvectors(
+        Function<MtermvectorsRequest.Builder, ObjectBuilder<MtermvectorsRequest>> fn
+    ) throws IOException, OpenSearchException {
         return mtermvectors(fn.apply(new MtermvectorsRequest.Builder()).build());
     }
 
@@ -1064,8 +1062,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public MtermvectorsResponse mtermvectors() throws IOException, OpenSearchException {
-        return this.transport.performRequest(
+    public CompletableFuture<MtermvectorsResponse> mtermvectors() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
             new MtermvectorsRequest.Builder().build(),
             MtermvectorsRequest._ENDPOINT,
             this.transportOptions
@@ -1079,8 +1077,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      *
      */
-    public BooleanResponse ping() throws IOException, OpenSearchException {
-        return this.transport.performRequest(PingRequest._INSTANCE, PingRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<BooleanResponse> ping() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(PingRequest._INSTANCE, PingRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: put_script
@@ -1091,14 +1089,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public PutScriptResponse putScript(PutScriptRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<PutScriptResponse> putScript(PutScriptRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<PutScriptRequest, PutScriptResponse, ErrorResponse> endpoint = (JsonEndpoint<
             PutScriptRequest,
             PutScriptResponse,
             ErrorResponse>) PutScriptRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1110,8 +1108,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final PutScriptResponse putScript(Function<PutScriptRequest.Builder, ObjectBuilder<PutScriptRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<PutScriptResponse> putScript(Function<PutScriptRequest.Builder, ObjectBuilder<PutScriptRequest>> fn)
+        throws IOException, OpenSearchException {
         return putScript(fn.apply(new PutScriptRequest.Builder()).build());
     }
 
@@ -1124,14 +1122,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public RankEvalResponse rankEval(RankEvalRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<RankEvalResponse> rankEval(RankEvalRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<RankEvalRequest, RankEvalResponse, ErrorResponse> endpoint = (JsonEndpoint<
             RankEvalRequest,
             RankEvalResponse,
             ErrorResponse>) RankEvalRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1144,8 +1142,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final RankEvalResponse rankEval(Function<RankEvalRequest.Builder, ObjectBuilder<RankEvalRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<RankEvalResponse> rankEval(Function<RankEvalRequest.Builder, ObjectBuilder<RankEvalRequest>> fn)
+        throws IOException, OpenSearchException {
         return rankEval(fn.apply(new RankEvalRequest.Builder()).build());
     }
 
@@ -1159,14 +1157,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public ReindexResponse reindex(ReindexRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<ReindexResponse> reindex(ReindexRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ReindexRequest, ReindexResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ReindexRequest,
             ReindexResponse,
             ErrorResponse>) ReindexRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1180,8 +1178,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final ReindexResponse reindex(Function<ReindexRequest.Builder, ObjectBuilder<ReindexRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<ReindexResponse> reindex(Function<ReindexRequest.Builder, ObjectBuilder<ReindexRequest>> fn)
+        throws IOException, OpenSearchException {
         return reindex(fn.apply(new ReindexRequest.Builder()).build());
     }
 
@@ -1193,8 +1191,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public ReindexResponse reindex() throws IOException, OpenSearchException {
-        return this.transport.performRequest(new ReindexRequest.Builder().build(), ReindexRequest._ENDPOINT, this.transportOptions);
+    public CompletableFuture<ReindexResponse> reindex() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(new ReindexRequest.Builder().build(), ReindexRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: reindex_rethrottle
@@ -1205,14 +1203,15 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public ReindexRethrottleResponse reindexRethrottle(ReindexRethrottleRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<ReindexRethrottleResponse> reindexRethrottle(ReindexRethrottleRequest request) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ReindexRethrottleRequest, ReindexRethrottleResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ReindexRethrottleRequest,
             ReindexRethrottleResponse,
             ErrorResponse>) ReindexRethrottleRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1224,7 +1223,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final ReindexRethrottleResponse reindexRethrottle(
+    public final CompletableFuture<ReindexRethrottleResponse> reindexRethrottle(
         Function<ReindexRethrottleRequest.Builder, ObjectBuilder<ReindexRethrottleRequest>> fn
     ) throws IOException, OpenSearchException {
         return reindexRethrottle(fn.apply(new ReindexRethrottleRequest.Builder()).build());
@@ -1238,14 +1237,15 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public RenderSearchTemplateResponse renderSearchTemplate(RenderSearchTemplateRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<RenderSearchTemplateResponse> renderSearchTemplate(RenderSearchTemplateRequest request) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<RenderSearchTemplateRequest, RenderSearchTemplateResponse, ErrorResponse> endpoint = (JsonEndpoint<
             RenderSearchTemplateRequest,
             RenderSearchTemplateResponse,
             ErrorResponse>) RenderSearchTemplateRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1257,7 +1257,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final RenderSearchTemplateResponse renderSearchTemplate(
+    public final CompletableFuture<RenderSearchTemplateResponse> renderSearchTemplate(
         Function<RenderSearchTemplateRequest.Builder, ObjectBuilder<RenderSearchTemplateRequest>> fn
     ) throws IOException, OpenSearchException {
         return renderSearchTemplate(fn.apply(new RenderSearchTemplateRequest.Builder()).build());
@@ -1269,8 +1269,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public RenderSearchTemplateResponse renderSearchTemplate() throws IOException, OpenSearchException {
-        return this.transport.performRequest(
+    public CompletableFuture<RenderSearchTemplateResponse> renderSearchTemplate() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
             new RenderSearchTemplateRequest.Builder().build(),
             RenderSearchTemplateRequest._ENDPOINT,
             this.transportOptions
@@ -1285,7 +1285,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TResult> ScriptsPainlessExecuteResponse<TResult> scriptsPainlessExecute(
+    public <TResult> CompletableFuture<ScriptsPainlessExecuteResponse<TResult>> scriptsPainlessExecute(
         ScriptsPainlessExecuteRequest request,
         Class<TResult> tResultClass
     ) throws IOException, OpenSearchException {
@@ -1300,7 +1300,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
             getDeserializer(tResultClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1312,7 +1312,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TResult> ScriptsPainlessExecuteResponse<TResult> scriptsPainlessExecute(
+    public final <TResult> CompletableFuture<ScriptsPainlessExecuteResponse<TResult>> scriptsPainlessExecute(
         Function<ScriptsPainlessExecuteRequest.Builder, ObjectBuilder<ScriptsPainlessExecuteRequest>> fn,
         Class<TResult> tResultClass
     ) throws IOException, OpenSearchException {
@@ -1326,8 +1326,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument> ScrollResponse<TDocument> scroll(ScrollRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<ScrollResponse<TDocument>> scroll(ScrollRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ScrollRequest, ScrollResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             ScrollRequest,
@@ -1339,7 +1339,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1350,7 +1350,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *            {@link ScrollRequest}
      */
 
-    public final <TDocument> ScrollResponse<TDocument> scroll(
+    public final <TDocument> CompletableFuture<ScrollResponse<TDocument>> scroll(
         Function<ScrollRequest.Builder, ObjectBuilder<ScrollRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1365,8 +1365,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument> SearchResponse<TDocument> search(SearchRequest request, Class<TDocument> tDocumentClass) throws IOException,
-        OpenSearchException {
+    public <TDocument> CompletableFuture<SearchResponse<TDocument>> search(SearchRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<SearchRequest, SearchResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             SearchRequest,
@@ -1378,7 +1378,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1390,7 +1390,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TDocument> SearchResponse<TDocument> search(
+    public final <TDocument> CompletableFuture<SearchResponse<TDocument>> search(
         Function<SearchRequest.Builder, ObjectBuilder<SearchRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1406,14 +1406,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public SearchShardsResponse searchShards(SearchShardsRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<SearchShardsResponse> searchShards(SearchShardsRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<SearchShardsRequest, SearchShardsResponse, ErrorResponse> endpoint = (JsonEndpoint<
             SearchShardsRequest,
             SearchShardsResponse,
             ErrorResponse>) SearchShardsRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1426,8 +1426,9 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final SearchShardsResponse searchShards(Function<SearchShardsRequest.Builder, ObjectBuilder<SearchShardsRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<SearchShardsResponse> searchShards(
+        Function<SearchShardsRequest.Builder, ObjectBuilder<SearchShardsRequest>> fn
+    ) throws IOException, OpenSearchException {
         return searchShards(fn.apply(new SearchShardsRequest.Builder()).build());
     }
 
@@ -1438,8 +1439,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public SearchShardsResponse searchShards() throws IOException, OpenSearchException {
-        return this.transport.performRequest(
+    public CompletableFuture<SearchShardsResponse> searchShards() throws IOException, OpenSearchException {
+        return this.transport.performRequestAsync(
             new SearchShardsRequest.Builder().build(),
             SearchShardsRequest._ENDPOINT,
             this.transportOptions
@@ -1454,8 +1455,10 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument> SearchTemplateResponse<TDocument> searchTemplate(SearchTemplateRequest request, Class<TDocument> tDocumentClass)
-        throws IOException, OpenSearchException {
+    public <TDocument> CompletableFuture<SearchTemplateResponse<TDocument>> searchTemplate(
+        SearchTemplateRequest request,
+        Class<TDocument> tDocumentClass
+    ) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<SearchTemplateRequest, SearchTemplateResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             SearchTemplateRequest,
@@ -1466,7 +1469,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
             "org.opensearch.client:Deserializer:_global.search_template.TDocument",
             getDeserializer(tDocumentClass)
         );
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1478,7 +1482,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TDocument> SearchTemplateResponse<TDocument> searchTemplate(
+    public final <TDocument> CompletableFuture<SearchTemplateResponse<TDocument>> searchTemplate(
         Function<SearchTemplateRequest.Builder, ObjectBuilder<SearchTemplateRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1495,14 +1499,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public TermsEnumResponse termsEnum(TermsEnumRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<TermsEnumResponse> termsEnum(TermsEnumRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<TermsEnumRequest, TermsEnumResponse, ErrorResponse> endpoint = (JsonEndpoint<
             TermsEnumRequest,
             TermsEnumResponse,
             ErrorResponse>) TermsEnumRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1516,8 +1520,8 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final TermsEnumResponse termsEnum(Function<TermsEnumRequest.Builder, ObjectBuilder<TermsEnumRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final CompletableFuture<TermsEnumResponse> termsEnum(Function<TermsEnumRequest.Builder, ObjectBuilder<TermsEnumRequest>> fn)
+        throws IOException, OpenSearchException {
         return termsEnum(fn.apply(new TermsEnumRequest.Builder()).build());
     }
 
@@ -1530,14 +1534,15 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument> TermvectorsResponse termvectors(TermvectorsRequest<TDocument> request) throws IOException, OpenSearchException {
+    public <TDocument> CompletableFuture<TermvectorsResponse> termvectors(TermvectorsRequest<TDocument> request) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<TermvectorsRequest<?>, TermvectorsResponse, ErrorResponse> endpoint = (JsonEndpoint<
             TermvectorsRequest<?>,
             TermvectorsResponse,
             ErrorResponse>) TermvectorsRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1550,7 +1555,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TDocument> TermvectorsResponse termvectors(
+    public final <TDocument> CompletableFuture<TermvectorsResponse> termvectors(
         Function<TermvectorsRequest.Builder<TDocument>, ObjectBuilder<TermvectorsRequest<TDocument>>> fn
     ) throws IOException, OpenSearchException {
         return termvectors(fn.apply(new TermvectorsRequest.Builder<TDocument>()).build());
@@ -1564,7 +1569,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public <TDocument, TPartialDocument> UpdateResponse<TDocument> update(
+    public <TDocument, TPartialDocument> CompletableFuture<UpdateResponse<TDocument>> update(
         UpdateRequest<TDocument, TPartialDocument> request,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1579,7 +1584,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1591,8 +1596,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final <TDocument, TPartialDocument> UpdateResponse<TDocument> update(
-
+    public final <TDocument, TPartialDocument> CompletableFuture<UpdateResponse<TDocument>> update(
         Function<UpdateRequest.Builder<TDocument, TPartialDocument>, ObjectBuilder<UpdateRequest<TDocument, TPartialDocument>>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1608,14 +1612,14 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public UpdateByQueryResponse updateByQuery(UpdateByQueryRequest request) throws IOException, OpenSearchException {
+    public CompletableFuture<UpdateByQueryResponse> updateByQuery(UpdateByQueryRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<UpdateByQueryRequest, UpdateByQueryResponse, ErrorResponse> endpoint = (JsonEndpoint<
             UpdateByQueryRequest,
             UpdateByQueryResponse,
             ErrorResponse>) UpdateByQueryRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1628,8 +1632,9 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final UpdateByQueryResponse updateByQuery(Function<UpdateByQueryRequest.Builder, ObjectBuilder<UpdateByQueryRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CompletableFuture<UpdateByQueryResponse> updateByQuery(
+        Function<UpdateByQueryRequest.Builder, ObjectBuilder<UpdateByQueryRequest>> fn
+    ) throws IOException, OpenSearchException {
         return updateByQuery(fn.apply(new UpdateByQueryRequest.Builder()).build());
     }
 
@@ -1642,15 +1647,15 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public UpdateByQueryRethrottleResponse updateByQueryRethrottle(UpdateByQueryRethrottleRequest request) throws IOException,
-        OpenSearchException {
+    public CompletableFuture<UpdateByQueryRethrottleResponse> updateByQueryRethrottle(UpdateByQueryRethrottleRequest request)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<UpdateByQueryRethrottleRequest, UpdateByQueryRethrottleResponse, ErrorResponse> endpoint = (JsonEndpoint<
             UpdateByQueryRethrottleRequest,
             UpdateByQueryRethrottleResponse,
             ErrorResponse>) UpdateByQueryRethrottleRequest._ENDPOINT;
 
-        return this.transport.performRequest(request, endpoint, this.transportOptions);
+        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1663,7 +1668,7 @@ public class OpenSearchClient extends ApiClient<OpenSearchTransport, OpenSearchC
      *
      */
 
-    public final UpdateByQueryRethrottleResponse updateByQueryRethrottle(
+    public final CompletableFuture<UpdateByQueryRethrottleResponse> updateByQueryRethrottle(
         Function<UpdateByQueryRethrottleRequest.Builder, ObjectBuilder<UpdateByQueryRethrottleRequest>> fn
     ) throws IOException, OpenSearchException {
         return updateByQueryRethrottle(fn.apply(new UpdateByQueryRethrottleRequest.Builder()).build());
