+++ java-client/src/main/java/org/opensearch/client/opensearch/OpenSearchClientBase.java
@@ -33,14 +33,13 @@
 package org.opensearch.client.opensearch;
 
 import java.io.IOException;
-import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 import javax.annotation.Nullable;
 import org.opensearch.client.ApiClient;
 import org.opensearch.client.opensearch._types.ErrorResponse;
 import org.opensearch.client.opensearch._types.OpenSearchException;
-import org.opensearch.client.opensearch.cat.OpenSearchCatAsyncClient;
-import org.opensearch.client.opensearch.cluster.OpenSearchClusterAsyncClient;
+import org.opensearch.client.opensearch.cat.OpenSearchCatClient;
+import org.opensearch.client.opensearch.cluster.OpenSearchClusterClient;
 import org.opensearch.client.opensearch.core.BulkRequest;
 import org.opensearch.client.opensearch.core.BulkResponse;
 import org.opensearch.client.opensearch.core.ClearScrollRequest;
@@ -122,14 +121,15 @@ import org.opensearch.client.opensearch.core.pit.DeletePitRequest;
 import org.opensearch.client.opensearch.core.pit.DeletePitResponse;
 import org.opensearch.client.opensearch.core.pit.ListAllPitRequest;
 import org.opensearch.client.opensearch.core.pit.ListAllPitResponse;
-import org.opensearch.client.opensearch.dangling_indices.OpenSearchDanglingIndicesAsyncClient;
-import org.opensearch.client.opensearch.features.OpenSearchFeaturesAsyncClient;
-import org.opensearch.client.opensearch.indices.OpenSearchIndicesAsyncClient;
-import org.opensearch.client.opensearch.ingest.OpenSearchIngestAsyncClient;
-import org.opensearch.client.opensearch.nodes.OpenSearchNodesAsyncClient;
-import org.opensearch.client.opensearch.shutdown.OpenSearchShutdownAsyncClient;
-import org.opensearch.client.opensearch.snapshot.OpenSearchSnapshotAsyncClient;
-import org.opensearch.client.opensearch.tasks.OpenSearchTasksAsyncClient;
+import org.opensearch.client.opensearch.dangling_indices.OpenSearchDanglingIndicesClient;
+import org.opensearch.client.opensearch.features.OpenSearchFeaturesClient;
+import org.opensearch.client.opensearch.generic.OpenSearchGenericClient;
+import org.opensearch.client.opensearch.indices.OpenSearchIndicesClient;
+import org.opensearch.client.opensearch.ingest.OpenSearchIngestClient;
+import org.opensearch.client.opensearch.nodes.OpenSearchNodesClient;
+import org.opensearch.client.opensearch.shutdown.OpenSearchShutdownClient;
+import org.opensearch.client.opensearch.snapshot.OpenSearchSnapshotClient;
+import org.opensearch.client.opensearch.tasks.OpenSearchTasksClient;
 import org.opensearch.client.transport.JsonEndpoint;
 import org.opensearch.client.transport.OpenSearchTransport;
 import org.opensearch.client.transport.TransportOptions;
@@ -140,61 +140,56 @@ import org.opensearch.client.util.ObjectBuilder;
 /**
  * Client for the namespace.
  */
-public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSearchAsyncClient> {
+public abstract class OpenSearchClientBase<Self extends OpenSearchClientBase<Self>>
+        extends ApiClient<OpenSearchTransport, Self> {
 
-    public OpenSearchAsyncClient(OpenSearchTransport transport) {
-        super(transport, null);
-    }
-
-    public OpenSearchAsyncClient(OpenSearchTransport transport, @Nullable TransportOptions transportOptions) {
+    public OpenSearchClientBase(OpenSearchTransport transport, @Nullable TransportOptions transportOptions) {
         super(transport, transportOptions);
     }
 
-    @Override
-    public OpenSearchAsyncClient withTransportOptions(@Nullable TransportOptions transportOptions) {
-        return new OpenSearchAsyncClient(this.transport, transportOptions);
-    }
-
     // ----- Child clients
-
-    public OpenSearchCatAsyncClient cat() {
-        return new OpenSearchCatAsyncClient(this.transport, this.transportOptions);
+    public OpenSearchGenericClient generic() {
+        return new OpenSearchGenericClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchClusterAsyncClient cluster() {
-        return new OpenSearchClusterAsyncClient(this.transport, this.transportOptions);
+    public OpenSearchCatClient cat() {
+        return new OpenSearchCatClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchDanglingIndicesAsyncClient danglingIndices() {
-        return new OpenSearchDanglingIndicesAsyncClient(this.transport, this.transportOptions);
+    public OpenSearchClusterClient cluster() {
+        return new OpenSearchClusterClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchFeaturesAsyncClient features() {
-        return new OpenSearchFeaturesAsyncClient(this.transport, this.transportOptions);
+    public OpenSearchDanglingIndicesClient danglingIndices() {
+        return new OpenSearchDanglingIndicesClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchIndicesAsyncClient indices() {
-        return new OpenSearchIndicesAsyncClient(this.transport, this.transportOptions);
+    public OpenSearchFeaturesClient features() {
+        return new OpenSearchFeaturesClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchIngestAsyncClient ingest() {
-        return new OpenSearchIngestAsyncClient(this.transport, this.transportOptions);
+    public OpenSearchIndicesClient indices() {
+        return new OpenSearchIndicesClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchNodesAsyncClient nodes() {
-        return new OpenSearchNodesAsyncClient(this.transport, this.transportOptions);
+    public OpenSearchIngestClient ingest() {
+        return new OpenSearchIngestClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchShutdownAsyncClient shutdown() {
-        return new OpenSearchShutdownAsyncClient(this.transport, this.transportOptions);
+    public OpenSearchNodesClient nodes() {
+        return new OpenSearchNodesClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchSnapshotAsyncClient snapshot() {
-        return new OpenSearchSnapshotAsyncClient(this.transport, this.transportOptions);
+    public OpenSearchShutdownClient shutdown() {
+        return new OpenSearchShutdownClient(this.transport, this.transportOptions);
     }
 
-    public OpenSearchTasksAsyncClient tasks() {
-        return new OpenSearchTasksAsyncClient(this.transport, this.transportOptions);
+    public OpenSearchSnapshotClient snapshot() {
+        return new OpenSearchSnapshotClient(this.transport, this.transportOptions);
+    }
+
+    public OpenSearchTasksClient tasks() {
+        return new OpenSearchTasksClient(this.transport, this.transportOptions);
     }
 
     // ----- Endpoint: bulk
@@ -206,14 +201,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<BulkResponse> bulk(BulkRequest request) throws IOException, OpenSearchException {
+    public BulkResponse bulk(BulkRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<BulkRequest, BulkResponse, ErrorResponse> endpoint = (JsonEndpoint<
             BulkRequest,
             BulkResponse,
             ErrorResponse>) BulkRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -226,8 +221,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<BulkResponse> bulk(Function<BulkRequest.Builder, ObjectBuilder<BulkRequest>> fn) throws IOException,
-        OpenSearchException {
+    public final BulkResponse bulk(Function<BulkRequest.Builder, ObjectBuilder<BulkRequest>> fn) throws IOException, OpenSearchException {
         return bulk(fn.apply(new BulkRequest.Builder()).build());
     }
 
@@ -238,8 +232,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<BulkResponse> bulk() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(new BulkRequest.Builder().build(), BulkRequest._ENDPOINT, this.transportOptions);
+    public BulkResponse bulk() throws IOException, OpenSearchException {
+        return this.transport.performRequest(new BulkRequest.Builder().build(), BulkRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: clear_scroll
@@ -250,14 +244,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<ClearScrollResponse> clearScroll(ClearScrollRequest request) throws IOException, OpenSearchException {
+    public ClearScrollResponse clearScroll(ClearScrollRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ClearScrollRequest, ClearScrollResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ClearScrollRequest,
             ClearScrollResponse,
             ErrorResponse>) ClearScrollRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -269,9 +263,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<ClearScrollResponse> clearScroll(
-        Function<ClearScrollRequest.Builder, ObjectBuilder<ClearScrollRequest>> fn
-    ) throws IOException, OpenSearchException {
+    public final ClearScrollResponse clearScroll(Function<ClearScrollRequest.Builder, ObjectBuilder<ClearScrollRequest>> fn)
+        throws IOException, OpenSearchException {
         return clearScroll(fn.apply(new ClearScrollRequest.Builder()).build());
     }
 
@@ -281,12 +274,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<ClearScrollResponse> clearScroll() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(
-            new ClearScrollRequest.Builder().build(),
-            ClearScrollRequest._ENDPOINT,
-            this.transportOptions
-        );
+    public ClearScrollResponse clearScroll() throws IOException, OpenSearchException {
+        return this.transport.performRequest(new ClearScrollRequest.Builder().build(), ClearScrollRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: count
@@ -297,14 +286,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<CountResponse> count(CountRequest request) throws IOException, OpenSearchException {
+    public CountResponse count(CountRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<CountRequest, CountResponse, ErrorResponse> endpoint = (JsonEndpoint<
             CountRequest,
             CountResponse,
             ErrorResponse>) CountRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -316,7 +305,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<CountResponse> count(Function<CountRequest.Builder, ObjectBuilder<CountRequest>> fn) throws IOException,
+    public final CountResponse count(Function<CountRequest.Builder, ObjectBuilder<CountRequest>> fn) throws IOException,
         OpenSearchException {
         return count(fn.apply(new CountRequest.Builder()).build());
     }
@@ -327,8 +316,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<CountResponse> count() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(new CountRequest.Builder().build(), CountRequest._ENDPOINT, this.transportOptions);
+    public CountResponse count() throws IOException, OpenSearchException {
+        return this.transport.performRequest(new CountRequest.Builder().build(), CountRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: create
@@ -342,14 +331,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument> CompletableFuture<CreateResponse> create(CreateRequest<TDocument> request) throws IOException, OpenSearchException {
+    public <TDocument> CreateResponse create(CreateRequest<TDocument> request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<CreateRequest<?>, CreateResponse, ErrorResponse> endpoint = (JsonEndpoint<
             CreateRequest<?>,
             CreateResponse,
             ErrorResponse>) CreateRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -364,9 +353,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TDocument> CompletableFuture<CreateResponse> create(
-        Function<CreateRequest.Builder<TDocument>, ObjectBuilder<CreateRequest<TDocument>>> fn
-    ) throws IOException, OpenSearchException {
+    public final <TDocument> CreateResponse create(Function<CreateRequest.Builder<TDocument>, ObjectBuilder<CreateRequest<TDocument>>> fn)
+        throws IOException, OpenSearchException {
         return create(fn.apply(new CreateRequest.Builder<TDocument>()).build());
     }
 
@@ -379,14 +367,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<CreatePitResponse> createPit(CreatePitRequest request) throws IOException, OpenSearchException {
+    public CreatePitResponse createPit(CreatePitRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<CreatePitRequest, CreatePitResponse, ErrorResponse> endpoint = (JsonEndpoint<
             CreatePitRequest,
             CreatePitResponse,
             ErrorResponse>) CreatePitRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -399,8 +387,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<CreatePitResponse> createPit(Function<CreatePitRequest.Builder, ObjectBuilder<CreatePitRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final CreatePitResponse createPit(Function<CreatePitRequest.Builder, ObjectBuilder<CreatePitRequest>> fn) throws IOException,
+        OpenSearchException {
         return createPit(fn.apply(new CreatePitRequest.Builder()).build());
     }
 
@@ -412,14 +400,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<DeleteResponse> delete(DeleteRequest request) throws IOException, OpenSearchException {
+    public DeleteResponse delete(DeleteRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeleteRequest, DeleteResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeleteRequest,
             DeleteResponse,
             ErrorResponse>) DeleteRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -431,8 +419,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<DeleteResponse> delete(Function<DeleteRequest.Builder, ObjectBuilder<DeleteRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final DeleteResponse delete(Function<DeleteRequest.Builder, ObjectBuilder<DeleteRequest>> fn) throws IOException,
+        OpenSearchException {
         return delete(fn.apply(new DeleteRequest.Builder()).build());
     }
 
@@ -444,14 +432,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<DeletePitResponse> deletePit(DeletePitRequest request) throws IOException, OpenSearchException {
+    public DeletePitResponse deletePit(DeletePitRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeletePitRequest, DeletePitResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeletePitRequest,
             DeletePitResponse,
             ErrorResponse>) DeletePitRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -463,8 +451,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<DeletePitResponse> deletePit(Function<DeletePitRequest.Builder, ObjectBuilder<DeletePitRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final DeletePitResponse deletePit(Function<DeletePitRequest.Builder, ObjectBuilder<DeletePitRequest>> fn) throws IOException,
+        OpenSearchException {
         return deletePit(fn.apply(new DeletePitRequest.Builder()).build());
     }
 
@@ -476,14 +464,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<DeleteByQueryResponse> deleteByQuery(DeleteByQueryRequest request) throws IOException, OpenSearchException {
+    public DeleteByQueryResponse deleteByQuery(DeleteByQueryRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeleteByQueryRequest, DeleteByQueryResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeleteByQueryRequest,
             DeleteByQueryResponse,
             ErrorResponse>) DeleteByQueryRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -495,9 +483,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<DeleteByQueryResponse> deleteByQuery(
-        Function<DeleteByQueryRequest.Builder, ObjectBuilder<DeleteByQueryRequest>> fn
-    ) throws IOException, OpenSearchException {
+    public final DeleteByQueryResponse deleteByQuery(Function<DeleteByQueryRequest.Builder, ObjectBuilder<DeleteByQueryRequest>> fn)
+        throws IOException, OpenSearchException {
         return deleteByQuery(fn.apply(new DeleteByQueryRequest.Builder()).build());
     }
 
@@ -510,15 +497,15 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<DeleteByQueryRethrottleResponse> deleteByQueryRethrottle(DeleteByQueryRethrottleRequest request)
-        throws IOException, OpenSearchException {
+    public DeleteByQueryRethrottleResponse deleteByQueryRethrottle(DeleteByQueryRethrottleRequest request) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeleteByQueryRethrottleRequest, DeleteByQueryRethrottleResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeleteByQueryRethrottleRequest,
             DeleteByQueryRethrottleResponse,
             ErrorResponse>) DeleteByQueryRethrottleRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -531,7 +518,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<DeleteByQueryRethrottleResponse> deleteByQueryRethrottle(
+    public final DeleteByQueryRethrottleResponse deleteByQueryRethrottle(
         Function<DeleteByQueryRethrottleRequest.Builder, ObjectBuilder<DeleteByQueryRethrottleRequest>> fn
     ) throws IOException, OpenSearchException {
         return deleteByQueryRethrottle(fn.apply(new DeleteByQueryRethrottleRequest.Builder()).build());
@@ -545,14 +532,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<DeleteScriptResponse> deleteScript(DeleteScriptRequest request) throws IOException, OpenSearchException {
+    public DeleteScriptResponse deleteScript(DeleteScriptRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<DeleteScriptRequest, DeleteScriptResponse, ErrorResponse> endpoint = (JsonEndpoint<
             DeleteScriptRequest,
             DeleteScriptResponse,
             ErrorResponse>) DeleteScriptRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -564,9 +551,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<DeleteScriptResponse> deleteScript(
-        Function<DeleteScriptRequest.Builder, ObjectBuilder<DeleteScriptRequest>> fn
-    ) throws IOException, OpenSearchException {
+    public final DeleteScriptResponse deleteScript(Function<DeleteScriptRequest.Builder, ObjectBuilder<DeleteScriptRequest>> fn)
+        throws IOException, OpenSearchException {
         return deleteScript(fn.apply(new DeleteScriptRequest.Builder()).build());
     }
 
@@ -578,14 +564,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<BooleanResponse> exists(ExistsRequest request) throws IOException, OpenSearchException {
+    public BooleanResponse exists(ExistsRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ExistsRequest, BooleanResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ExistsRequest,
             BooleanResponse,
             ErrorResponse>) ExistsRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -597,8 +583,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<BooleanResponse> exists(Function<ExistsRequest.Builder, ObjectBuilder<ExistsRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final BooleanResponse exists(Function<ExistsRequest.Builder, ObjectBuilder<ExistsRequest>> fn) throws IOException,
+        OpenSearchException {
         return exists(fn.apply(new ExistsRequest.Builder()).build());
     }
 
@@ -610,14 +596,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<BooleanResponse> existsSource(ExistsSourceRequest request) throws IOException, OpenSearchException {
+    public BooleanResponse existsSource(ExistsSourceRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ExistsSourceRequest, BooleanResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ExistsSourceRequest,
             BooleanResponse,
             ErrorResponse>) ExistsSourceRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -629,9 +615,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<BooleanResponse> existsSource(
-        Function<ExistsSourceRequest.Builder, ObjectBuilder<ExistsSourceRequest>> fn
-    ) throws IOException, OpenSearchException {
+    public final BooleanResponse existsSource(Function<ExistsSourceRequest.Builder, ObjectBuilder<ExistsSourceRequest>> fn)
+        throws IOException, OpenSearchException {
         return existsSource(fn.apply(new ExistsSourceRequest.Builder()).build());
     }
 
@@ -643,8 +628,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument> CompletableFuture<ExplainResponse<TDocument>> explain(ExplainRequest request, Class<TDocument> tDocumentClass)
-        throws IOException, OpenSearchException {
+    public <TDocument> ExplainResponse<TDocument> explain(ExplainRequest request, Class<TDocument> tDocumentClass) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ExplainRequest, ExplainResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             ExplainRequest,
@@ -656,7 +641,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -668,7 +653,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TDocument> CompletableFuture<ExplainResponse<TDocument>> explain(
+    public final <TDocument> ExplainResponse<TDocument> explain(
         Function<ExplainRequest.Builder, ObjectBuilder<ExplainRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -680,16 +665,18 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
     /**
      * Returns the information about the capabilities of fields among multiple
      * indices.
+     *
+     *
      */
 
-    public CompletableFuture<FieldCapsResponse> fieldCaps(FieldCapsRequest request) throws IOException, OpenSearchException {
+    public FieldCapsResponse fieldCaps(FieldCapsRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<FieldCapsRequest, FieldCapsResponse, ErrorResponse> endpoint = (JsonEndpoint<
             FieldCapsRequest,
             FieldCapsResponse,
             ErrorResponse>) FieldCapsRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -702,8 +689,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<FieldCapsResponse> fieldCaps(Function<FieldCapsRequest.Builder, ObjectBuilder<FieldCapsRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final FieldCapsResponse fieldCaps(Function<FieldCapsRequest.Builder, ObjectBuilder<FieldCapsRequest>> fn) throws IOException,
+        OpenSearchException {
         return fieldCaps(fn.apply(new FieldCapsRequest.Builder()).build());
     }
 
@@ -714,12 +701,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<FieldCapsResponse> fieldCaps() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(
-            new FieldCapsRequest.Builder().build(),
-            FieldCapsRequest._ENDPOINT,
-            this.transportOptions
-        );
+    public FieldCapsResponse fieldCaps() throws IOException, OpenSearchException {
+        return this.transport.performRequest(new FieldCapsRequest.Builder().build(), FieldCapsRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: get
@@ -730,8 +713,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument> CompletableFuture<GetResponse<TDocument>> get(GetRequest request, Class<TDocument> tDocumentClass)
-        throws IOException, OpenSearchException {
+    public <TDocument> GetResponse<TDocument> get(GetRequest request, Class<TDocument> tDocumentClass) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<GetRequest, GetResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             GetRequest,
@@ -743,7 +726,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -755,7 +738,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TDocument> CompletableFuture<GetResponse<TDocument>> get(
+    public final <TDocument> GetResponse<TDocument> get(
         Function<GetRequest.Builder, ObjectBuilder<GetRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -770,14 +753,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<GetScriptResponse> getScript(GetScriptRequest request) throws IOException, OpenSearchException {
+    public GetScriptResponse getScript(GetScriptRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<GetScriptRequest, GetScriptResponse, ErrorResponse> endpoint = (JsonEndpoint<
             GetScriptRequest,
             GetScriptResponse,
             ErrorResponse>) GetScriptRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -789,8 +772,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<GetScriptResponse> getScript(Function<GetScriptRequest.Builder, ObjectBuilder<GetScriptRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final GetScriptResponse getScript(Function<GetScriptRequest.Builder, ObjectBuilder<GetScriptRequest>> fn) throws IOException,
+        OpenSearchException {
         return getScript(fn.apply(new GetScriptRequest.Builder()).build());
     }
 
@@ -801,12 +784,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      *
      */
-    public CompletableFuture<GetScriptContextResponse> getScriptContext() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(
-            GetScriptContextRequest._INSTANCE,
-            GetScriptContextRequest._ENDPOINT,
-            this.transportOptions
-        );
+    public GetScriptContextResponse getScriptContext() throws IOException, OpenSearchException {
+        return this.transport.performRequest(GetScriptContextRequest._INSTANCE, GetScriptContextRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: get_script_languages
@@ -816,8 +795,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      *
      */
-    public CompletableFuture<GetScriptLanguagesResponse> getScriptLanguages() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(
+    public GetScriptLanguagesResponse getScriptLanguages() throws IOException, OpenSearchException {
+        return this.transport.performRequest(
             GetScriptLanguagesRequest._INSTANCE,
             GetScriptLanguagesRequest._ENDPOINT,
             this.transportOptions
@@ -832,8 +811,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument> CompletableFuture<GetSourceResponse<TDocument>> getSource(GetSourceRequest request, Class<TDocument> tDocumentClass)
-        throws IOException, OpenSearchException {
+    public <TDocument> GetSourceResponse<TDocument> getSource(GetSourceRequest request, Class<TDocument> tDocumentClass) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<GetSourceRequest, GetSourceResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             GetSourceRequest,
@@ -845,7 +824,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -857,7 +836,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TDocument> CompletableFuture<GetSourceResponse<TDocument>> getSource(
+    public final <TDocument> GetSourceResponse<TDocument> getSource(
         Function<GetSourceRequest.Builder, ObjectBuilder<GetSourceRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -872,14 +851,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument> CompletableFuture<IndexResponse> index(IndexRequest<TDocument> request) throws IOException, OpenSearchException {
+    public <TDocument> IndexResponse index(IndexRequest<TDocument> request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<IndexRequest<?>, IndexResponse, ErrorResponse> endpoint = (JsonEndpoint<
             IndexRequest<?>,
             IndexResponse,
             ErrorResponse>) IndexRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -891,21 +870,9 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TDocument> CompletableFuture<IndexResponse> index(
-        Function<IndexRequest.Builder<TDocument>, ObjectBuilder<IndexRequest<TDocument>>> fn
-    ) throws IOException, OpenSearchException {
+    public final <TDocument> IndexResponse index(Function<IndexRequest.Builder<TDocument>, ObjectBuilder<IndexRequest<TDocument>>> fn)
+        throws IOException, OpenSearchException {
         return index(fn.apply(new IndexRequest.Builder<TDocument>()).build());
-    }
-
-    // ----- Endpoint: info
-
-    /**
-     * Returns basic information about the cluster.
-     *
-     *
-     */
-    public CompletableFuture<InfoResponse> info() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(InfoRequest._INSTANCE, InfoRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: list_point_in_time
@@ -916,8 +883,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<ListAllPitResponse> listAllPit() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(ListAllPitRequest._INSTANCE, ListAllPitRequest._ENDPOINT, this.transportOptions);
+    public ListAllPitResponse listAllPit() throws IOException, OpenSearchException {
+        return this.transport.performRequest(ListAllPitRequest._INSTANCE, ListAllPitRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: mget
@@ -928,8 +895,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument> CompletableFuture<MgetResponse<TDocument>> mget(MgetRequest request, Class<TDocument> tDocumentClass)
-        throws IOException, OpenSearchException {
+    public <TDocument> MgetResponse<TDocument> mget(MgetRequest request, Class<TDocument> tDocumentClass) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<MgetRequest, MgetResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             MgetRequest,
@@ -941,7 +908,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -953,7 +920,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TDocument> CompletableFuture<MgetResponse<TDocument>> mget(
+    public final <TDocument> MgetResponse<TDocument> mget(
         Function<MgetRequest.Builder, ObjectBuilder<MgetRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -968,8 +935,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument> CompletableFuture<MsearchResponse<TDocument>> msearch(MsearchRequest request, Class<TDocument> tDocumentClass)
-        throws IOException, OpenSearchException {
+    public <TDocument> MsearchResponse<TDocument> msearch(MsearchRequest request, Class<TDocument> tDocumentClass) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<MsearchRequest, MsearchResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             MsearchRequest,
@@ -981,7 +948,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -993,7 +960,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TDocument> CompletableFuture<MsearchResponse<TDocument>> msearch(
+    public final <TDocument> MsearchResponse<TDocument> msearch(
         Function<MsearchRequest.Builder, ObjectBuilder<MsearchRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1008,10 +975,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument> CompletableFuture<MsearchTemplateResponse<TDocument>> msearchTemplate(
-        MsearchTemplateRequest request,
-        Class<TDocument> tDocumentClass
-    ) throws IOException, OpenSearchException {
+    public <TDocument> MsearchTemplateResponse<TDocument> msearchTemplate(MsearchTemplateRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<MsearchTemplateRequest, MsearchTemplateResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             MsearchTemplateRequest,
@@ -1023,7 +988,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1035,7 +1000,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TDocument> CompletableFuture<MsearchTemplateResponse<TDocument>> msearchTemplate(
+    public final <TDocument> MsearchTemplateResponse<TDocument> msearchTemplate(
         Function<MsearchTemplateRequest.Builder, ObjectBuilder<MsearchTemplateRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1050,14 +1015,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<MtermvectorsResponse> mtermvectors(MtermvectorsRequest request) throws IOException, OpenSearchException {
+    public MtermvectorsResponse mtermvectors(MtermvectorsRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<MtermvectorsRequest, MtermvectorsResponse, ErrorResponse> endpoint = (JsonEndpoint<
             MtermvectorsRequest,
             MtermvectorsResponse,
             ErrorResponse>) MtermvectorsRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1069,9 +1034,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<MtermvectorsResponse> mtermvectors(
-        Function<MtermvectorsRequest.Builder, ObjectBuilder<MtermvectorsRequest>> fn
-    ) throws IOException, OpenSearchException {
+    public final MtermvectorsResponse mtermvectors(Function<MtermvectorsRequest.Builder, ObjectBuilder<MtermvectorsRequest>> fn)
+        throws IOException, OpenSearchException {
         return mtermvectors(fn.apply(new MtermvectorsRequest.Builder()).build());
     }
 
@@ -1081,8 +1045,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<MtermvectorsResponse> mtermvectors() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(
+    public MtermvectorsResponse mtermvectors() throws IOException, OpenSearchException {
+        return this.transport.performRequest(
             new MtermvectorsRequest.Builder().build(),
             MtermvectorsRequest._ENDPOINT,
             this.transportOptions
@@ -1096,8 +1060,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      *
      */
-    public CompletableFuture<BooleanResponse> ping() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(PingRequest._INSTANCE, PingRequest._ENDPOINT, this.transportOptions);
+    public BooleanResponse ping() throws IOException, OpenSearchException {
+        return this.transport.performRequest(PingRequest._INSTANCE, PingRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: put_script
@@ -1108,14 +1072,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<PutScriptResponse> putScript(PutScriptRequest request) throws IOException, OpenSearchException {
+    public PutScriptResponse putScript(PutScriptRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<PutScriptRequest, PutScriptResponse, ErrorResponse> endpoint = (JsonEndpoint<
             PutScriptRequest,
             PutScriptResponse,
             ErrorResponse>) PutScriptRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1127,8 +1091,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<PutScriptResponse> putScript(Function<PutScriptRequest.Builder, ObjectBuilder<PutScriptRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final PutScriptResponse putScript(Function<PutScriptRequest.Builder, ObjectBuilder<PutScriptRequest>> fn) throws IOException,
+        OpenSearchException {
         return putScript(fn.apply(new PutScriptRequest.Builder()).build());
     }
 
@@ -1141,14 +1105,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<RankEvalResponse> rankEval(RankEvalRequest request) throws IOException, OpenSearchException {
+    public RankEvalResponse rankEval(RankEvalRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<RankEvalRequest, RankEvalResponse, ErrorResponse> endpoint = (JsonEndpoint<
             RankEvalRequest,
             RankEvalResponse,
             ErrorResponse>) RankEvalRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1161,8 +1125,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<RankEvalResponse> rankEval(Function<RankEvalRequest.Builder, ObjectBuilder<RankEvalRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final RankEvalResponse rankEval(Function<RankEvalRequest.Builder, ObjectBuilder<RankEvalRequest>> fn) throws IOException,
+        OpenSearchException {
         return rankEval(fn.apply(new RankEvalRequest.Builder()).build());
     }
 
@@ -1176,14 +1140,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<ReindexResponse> reindex(ReindexRequest request) throws IOException, OpenSearchException {
+    public ReindexResponse reindex(ReindexRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ReindexRequest, ReindexResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ReindexRequest,
             ReindexResponse,
             ErrorResponse>) ReindexRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1197,8 +1161,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<ReindexResponse> reindex(Function<ReindexRequest.Builder, ObjectBuilder<ReindexRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final ReindexResponse reindex(Function<ReindexRequest.Builder, ObjectBuilder<ReindexRequest>> fn) throws IOException,
+        OpenSearchException {
         return reindex(fn.apply(new ReindexRequest.Builder()).build());
     }
 
@@ -1210,8 +1174,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<ReindexResponse> reindex() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(new ReindexRequest.Builder().build(), ReindexRequest._ENDPOINT, this.transportOptions);
+    public ReindexResponse reindex() throws IOException, OpenSearchException {
+        return this.transport.performRequest(new ReindexRequest.Builder().build(), ReindexRequest._ENDPOINT, this.transportOptions);
     }
 
     // ----- Endpoint: reindex_rethrottle
@@ -1222,15 +1186,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<ReindexRethrottleResponse> reindexRethrottle(ReindexRethrottleRequest request) throws IOException,
-        OpenSearchException {
+    public ReindexRethrottleResponse reindexRethrottle(ReindexRethrottleRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ReindexRethrottleRequest, ReindexRethrottleResponse, ErrorResponse> endpoint = (JsonEndpoint<
             ReindexRethrottleRequest,
             ReindexRethrottleResponse,
             ErrorResponse>) ReindexRethrottleRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1242,7 +1205,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<ReindexRethrottleResponse> reindexRethrottle(
+    public final ReindexRethrottleResponse reindexRethrottle(
         Function<ReindexRethrottleRequest.Builder, ObjectBuilder<ReindexRethrottleRequest>> fn
     ) throws IOException, OpenSearchException {
         return reindexRethrottle(fn.apply(new ReindexRethrottleRequest.Builder()).build());
@@ -1256,15 +1219,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<RenderSearchTemplateResponse> renderSearchTemplate(RenderSearchTemplateRequest request) throws IOException,
-        OpenSearchException {
+    public RenderSearchTemplateResponse renderSearchTemplate(RenderSearchTemplateRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<RenderSearchTemplateRequest, RenderSearchTemplateResponse, ErrorResponse> endpoint = (JsonEndpoint<
             RenderSearchTemplateRequest,
             RenderSearchTemplateResponse,
             ErrorResponse>) RenderSearchTemplateRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1276,7 +1238,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<RenderSearchTemplateResponse> renderSearchTemplate(
+    public final RenderSearchTemplateResponse renderSearchTemplate(
         Function<RenderSearchTemplateRequest.Builder, ObjectBuilder<RenderSearchTemplateRequest>> fn
     ) throws IOException, OpenSearchException {
         return renderSearchTemplate(fn.apply(new RenderSearchTemplateRequest.Builder()).build());
@@ -1288,8 +1250,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<RenderSearchTemplateResponse> renderSearchTemplate() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(
+    public RenderSearchTemplateResponse renderSearchTemplate() throws IOException, OpenSearchException {
+        return this.transport.performRequest(
             new RenderSearchTemplateRequest.Builder().build(),
             RenderSearchTemplateRequest._ENDPOINT,
             this.transportOptions
@@ -1304,7 +1266,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TResult> CompletableFuture<ScriptsPainlessExecuteResponse<TResult>> scriptsPainlessExecute(
+    public <TResult> ScriptsPainlessExecuteResponse<TResult> scriptsPainlessExecute(
         ScriptsPainlessExecuteRequest request,
         Class<TResult> tResultClass
     ) throws IOException, OpenSearchException {
@@ -1319,7 +1281,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
             getDeserializer(tResultClass)
         );
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1331,7 +1293,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TResult> CompletableFuture<ScriptsPainlessExecuteResponse<TResult>> scriptsPainlessExecute(
+    public final <TResult> ScriptsPainlessExecuteResponse<TResult> scriptsPainlessExecute(
         Function<ScriptsPainlessExecuteRequest.Builder, ObjectBuilder<ScriptsPainlessExecuteRequest>> fn,
         Class<TResult> tResultClass
     ) throws IOException, OpenSearchException {
@@ -1345,8 +1307,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument> CompletableFuture<ScrollResponse<TDocument>> scroll(ScrollRequest request, Class<TDocument> tDocumentClass)
-        throws IOException, OpenSearchException {
+    public <TDocument> ScrollResponse<TDocument> scroll(ScrollRequest request, Class<TDocument> tDocumentClass) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<ScrollRequest, ScrollResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             ScrollRequest,
@@ -1358,7 +1320,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1369,7 +1331,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *            {@link ScrollRequest}
      */
 
-    public final <TDocument> CompletableFuture<ScrollResponse<TDocument>> scroll(
+    public final <TDocument> ScrollResponse<TDocument> scroll(
         Function<ScrollRequest.Builder, ObjectBuilder<ScrollRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1384,8 +1346,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument> CompletableFuture<SearchResponse<TDocument>> search(SearchRequest request, Class<TDocument> tDocumentClass)
-        throws IOException, OpenSearchException {
+    public <TDocument> SearchResponse<TDocument> search(SearchRequest request, Class<TDocument> tDocumentClass) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<SearchRequest, SearchResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             SearchRequest,
@@ -1397,7 +1359,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1409,7 +1371,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TDocument> CompletableFuture<SearchResponse<TDocument>> search(
+    public final <TDocument> SearchResponse<TDocument> search(
         Function<SearchRequest.Builder, ObjectBuilder<SearchRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1425,14 +1387,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<SearchShardsResponse> searchShards(SearchShardsRequest request) throws IOException, OpenSearchException {
+    public SearchShardsResponse searchShards(SearchShardsRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<SearchShardsRequest, SearchShardsResponse, ErrorResponse> endpoint = (JsonEndpoint<
             SearchShardsRequest,
             SearchShardsResponse,
             ErrorResponse>) SearchShardsRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1445,9 +1407,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<SearchShardsResponse> searchShards(
-        Function<SearchShardsRequest.Builder, ObjectBuilder<SearchShardsRequest>> fn
-    ) throws IOException, OpenSearchException {
+    public final SearchShardsResponse searchShards(Function<SearchShardsRequest.Builder, ObjectBuilder<SearchShardsRequest>> fn)
+        throws IOException, OpenSearchException {
         return searchShards(fn.apply(new SearchShardsRequest.Builder()).build());
     }
 
@@ -1458,8 +1419,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<SearchShardsResponse> searchShards() throws IOException, OpenSearchException {
-        return this.transport.performRequestAsync(
+    public SearchShardsResponse searchShards() throws IOException, OpenSearchException {
+        return this.transport.performRequest(
             new SearchShardsRequest.Builder().build(),
             SearchShardsRequest._ENDPOINT,
             this.transportOptions
@@ -1474,10 +1435,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument> CompletableFuture<SearchTemplateResponse<TDocument>> searchTemplate(
-        SearchTemplateRequest request,
-        Class<TDocument> tDocumentClass
-    ) throws IOException, OpenSearchException {
+    public <TDocument> SearchTemplateResponse<TDocument> searchTemplate(SearchTemplateRequest request, Class<TDocument> tDocumentClass)
+        throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<SearchTemplateRequest, SearchTemplateResponse<TDocument>, ErrorResponse> endpoint = (JsonEndpoint<
             SearchTemplateRequest,
@@ -1488,8 +1447,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
             "org.opensearch.client:Deserializer:_global.search_template.TDocument",
             getDeserializer(tDocumentClass)
         );
-
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1501,7 +1459,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TDocument> CompletableFuture<SearchTemplateResponse<TDocument>> searchTemplate(
+    public final <TDocument> SearchTemplateResponse<TDocument> searchTemplate(
         Function<SearchTemplateRequest.Builder, ObjectBuilder<SearchTemplateRequest>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1518,14 +1476,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<TermsEnumResponse> termsEnum(TermsEnumRequest request) throws IOException, OpenSearchException {
+    public TermsEnumResponse termsEnum(TermsEnumRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<TermsEnumRequest, TermsEnumResponse, ErrorResponse> endpoint = (JsonEndpoint<
             TermsEnumRequest,
             TermsEnumResponse,
             ErrorResponse>) TermsEnumRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1539,8 +1497,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<TermsEnumResponse> termsEnum(Function<TermsEnumRequest.Builder, ObjectBuilder<TermsEnumRequest>> fn)
-        throws IOException, OpenSearchException {
+    public final TermsEnumResponse termsEnum(Function<TermsEnumRequest.Builder, ObjectBuilder<TermsEnumRequest>> fn) throws IOException,
+        OpenSearchException {
         return termsEnum(fn.apply(new TermsEnumRequest.Builder()).build());
     }
 
@@ -1553,15 +1511,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument> CompletableFuture<TermvectorsResponse> termvectors(TermvectorsRequest<TDocument> request) throws IOException,
-        OpenSearchException {
+    public <TDocument> TermvectorsResponse termvectors(TermvectorsRequest<TDocument> request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<TermvectorsRequest<?>, TermvectorsResponse, ErrorResponse> endpoint = (JsonEndpoint<
             TermvectorsRequest<?>,
             TermvectorsResponse,
             ErrorResponse>) TermvectorsRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1574,7 +1531,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TDocument> CompletableFuture<TermvectorsResponse> termvectors(
+    public final <TDocument> TermvectorsResponse termvectors(
         Function<TermvectorsRequest.Builder<TDocument>, ObjectBuilder<TermvectorsRequest<TDocument>>> fn
     ) throws IOException, OpenSearchException {
         return termvectors(fn.apply(new TermvectorsRequest.Builder<TDocument>()).build());
@@ -1588,7 +1545,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public <TDocument, TPartialDocument> CompletableFuture<UpdateResponse<TDocument>> update(
+    public <TDocument, TPartialDocument> UpdateResponse<TDocument> update(
         UpdateRequest<TDocument, TPartialDocument> request,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1603,7 +1560,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
             getDeserializer(tDocumentClass)
         );
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1615,7 +1572,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final <TDocument, TPartialDocument> CompletableFuture<UpdateResponse<TDocument>> update(
+    public final <TDocument, TPartialDocument> UpdateResponse<TDocument> update(
+
         Function<UpdateRequest.Builder<TDocument, TPartialDocument>, ObjectBuilder<UpdateRequest<TDocument, TPartialDocument>>> fn,
         Class<TDocument> tDocumentClass
     ) throws IOException, OpenSearchException {
@@ -1631,14 +1589,14 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<UpdateByQueryResponse> updateByQuery(UpdateByQueryRequest request) throws IOException, OpenSearchException {
+    public UpdateByQueryResponse updateByQuery(UpdateByQueryRequest request) throws IOException, OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<UpdateByQueryRequest, UpdateByQueryResponse, ErrorResponse> endpoint = (JsonEndpoint<
             UpdateByQueryRequest,
             UpdateByQueryResponse,
             ErrorResponse>) UpdateByQueryRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1651,9 +1609,8 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<UpdateByQueryResponse> updateByQuery(
-        Function<UpdateByQueryRequest.Builder, ObjectBuilder<UpdateByQueryRequest>> fn
-    ) throws IOException, OpenSearchException {
+    public final UpdateByQueryResponse updateByQuery(Function<UpdateByQueryRequest.Builder, ObjectBuilder<UpdateByQueryRequest>> fn)
+        throws IOException, OpenSearchException {
         return updateByQuery(fn.apply(new UpdateByQueryRequest.Builder()).build());
     }
 
@@ -1666,15 +1623,15 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public CompletableFuture<UpdateByQueryRethrottleResponse> updateByQueryRethrottle(UpdateByQueryRethrottleRequest request)
-        throws IOException, OpenSearchException {
+    public UpdateByQueryRethrottleResponse updateByQueryRethrottle(UpdateByQueryRethrottleRequest request) throws IOException,
+        OpenSearchException {
         @SuppressWarnings("unchecked")
         JsonEndpoint<UpdateByQueryRethrottleRequest, UpdateByQueryRethrottleResponse, ErrorResponse> endpoint = (JsonEndpoint<
             UpdateByQueryRethrottleRequest,
             UpdateByQueryRethrottleResponse,
             ErrorResponse>) UpdateByQueryRethrottleRequest._ENDPOINT;
 
-        return this.transport.performRequestAsync(request, endpoint, this.transportOptions);
+        return this.transport.performRequest(request, endpoint, this.transportOptions);
     }
 
     /**
@@ -1687,7 +1644,7 @@ public class OpenSearchAsyncClient extends ApiClient<OpenSearchTransport, OpenSe
      *
      */
 
-    public final CompletableFuture<UpdateByQueryRethrottleResponse> updateByQueryRethrottle(
+    public final UpdateByQueryRethrottleResponse updateByQueryRethrottle(
         Function<UpdateByQueryRethrottleRequest.Builder, ObjectBuilder<UpdateByQueryRethrottleRequest>> fn
     ) throws IOException, OpenSearchException {
         return updateByQueryRethrottle(fn.apply(new UpdateByQueryRethrottleRequest.Builder()).build());
